refactor(components): move code out of index and use it for export
refactor(pluginId): move into common constant
