+++ admin/src/components/PluginIcon/PluginIcon.js
@@ -1,0 +1,14 @@
+/**
+ *
+ * PluginIcon
+ *
+ */
+
+import React from 'react';
+import Play from '@strapi/icons/Play';
+
+const PluginIcon = () => {
+	return <Play />;
+};
+
+export default PluginIcon;
