+++ admin/src/components/PluginIcon/index.js
@@ -1,12 +1,1 @@
-/**
- *
- * PluginIcon
- *
- */
-
-import React from 'react';
-import Play from '@strapi/icons/Play';
-
-const PluginIcon = () => <Play />;
-
-export { PluginIcon };
+export { default } from './PluginIcon';
