--- admin/src/pluginId.js
@@ -1,3 +1,0 @@
-import pluginPkg from '../../package.json';
-
-export const pluginId = pluginPkg.strapi.name;
