+++ admin/src/utils/constants.js
@@ -1,0 +1,1 @@
+export const PLUGIN_ID = 'website-builder';
